requirements.txt
fastapi==0.115.4
pydantic==2.9.2
torch==2.5.1
python-dotenv==1.0.1
streamlit==1.39.0
requests==2.32.3